% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProvAbstractObject.R
\name{getProvClasses}
\alias{getProvClasses}
\title{getProvClasses}
\usage{
getProvClasses()
}
\description{
get the list of PROV classes, ie classes extending \link{ProvAbstractObject} super class,
including classes eventually defined outside \pkg{prov4R}. In case the latter is on the search path,
the list of PROV classes will be cached for optimized used by \pkg{prov4R} encoder/decoder.
}
\examples{
  getProvClasses()

}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
