% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProvNamespace.R
\name{getProvNamespace}
\alias{getProvNamespace}
\title{getProvNamespace}
\usage{
getProvNamespace(id)
}
\arguments{
\item{id}{namespace prefix}
}
\description{
\code{getProvNamespace} gets a namespace given its id
}
\examples{
  getProvNamespace("GMD")

}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
