% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProvNamespace.R
\name{getProvNamespaces}
\alias{getProvNamespaces}
\title{getProvNamespaces}
\usage{
getProvNamespaces()
}
\description{
\code{getProvNamespaces} gets the list of namespaces registered
}
\examples{
  getProvNamespaces()

}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
