% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProvSchemas.R
\name{getProvSchemas}
\alias{getProvSchemas}
\title{getProvSchemas}
\usage{
getProvSchemas()
}
\description{
\code{getProvSchemas} gets the schemas registered in \pkg{prov4R}
}
\examples{
  getProvSchemas()

}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
