% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prov4R.R
\docType{package}
\name{prov4R}
\alias{prov4R}
\alias{prov4R-package}
\title{Tools to Handle Metadata as PROV XML Format}
\description{
Provides tools to read/write metadata based on the PROV XML format.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/eblondel/prov4R}
  \item \url{https://www.w3.org/TR/prov-xml/}
  \item Report bugs at \url{https://github.com/eblondel/prov4R/issues}
}

}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
