% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prov4R_logger.R
\docType{class}
\name{prov4RLogger}
\alias{prov4RLogger}
\title{prov4RLogger}
\format{
\code{ \link[R6]{R6Class}} object.
}
\value{
Object of \code{ \link[R6]{R6Class}} for modelling a simple logger
}
\description{
prov4RLogger

prov4RLogger
}
\note{
Logger class used internally by prov4R
}
\keyword{logger}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{verbose.info}}{If package info log messages have to be printed out}

\item{\code{verbose.debug}}{If curl debug log messages have to be printed out}

\item{\code{loggerType}}{the type of logger}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-prov4RLogger-logger}{\code{prov4RLogger$logger()}}
\item \href{#method-prov4RLogger-INFO}{\code{prov4RLogger$INFO()}}
\item \href{#method-prov4RLogger-WARN}{\code{prov4RLogger$WARN()}}
\item \href{#method-prov4RLogger-ERROR}{\code{prov4RLogger$ERROR()}}
\item \href{#method-prov4RLogger-new}{\code{prov4RLogger$new()}}
\item \href{#method-prov4RLogger-getClassName}{\code{prov4RLogger$getClassName()}}
\item \href{#method-prov4RLogger-getClass}{\code{prov4RLogger$getClass()}}
\item \href{#method-prov4RLogger-clone}{\code{prov4RLogger$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prov4RLogger-logger"></a>}}
\if{latex}{\out{\hypertarget{method-prov4RLogger-logger}{}}}
\subsection{Method \code{logger()}}{
Provides log messages
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prov4RLogger$logger(type, text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{type of log ("INFO", "WARN", "ERROR")}

\item{\code{text}}{the log message text}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prov4RLogger-INFO"></a>}}
\if{latex}{\out{\hypertarget{method-prov4RLogger-INFO}{}}}
\subsection{Method \code{INFO()}}{
Provides INFO log messages
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prov4RLogger$INFO(text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{the log message text}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prov4RLogger-WARN"></a>}}
\if{latex}{\out{\hypertarget{method-prov4RLogger-WARN}{}}}
\subsection{Method \code{WARN()}}{
Provides WARN log messages
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prov4RLogger$WARN(text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{the log message text}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prov4RLogger-ERROR"></a>}}
\if{latex}{\out{\hypertarget{method-prov4RLogger-ERROR}{}}}
\subsection{Method \code{ERROR()}}{
Provides ERROR log messages
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prov4RLogger$ERROR(text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{the log message text}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prov4RLogger-new"></a>}}
\if{latex}{\out{\hypertarget{method-prov4RLogger-new}{}}}
\subsection{Method \code{new()}}{
Initalizes the logger
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prov4RLogger$new(logger = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{logger}}{logger type "INFO", "DEBUG" or \code{NULL}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prov4RLogger-getClassName"></a>}}
\if{latex}{\out{\hypertarget{method-prov4RLogger-getClassName}{}}}
\subsection{Method \code{getClassName()}}{
Get class name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prov4RLogger$getClassName()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
object of class \code{data.frame}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prov4RLogger-getClass"></a>}}
\if{latex}{\out{\hypertarget{method-prov4RLogger-getClass}{}}}
\subsection{Method \code{getClass()}}{
Get class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prov4RLogger$getClass()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
object of class  \link[R6]{R6Class}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prov4RLogger-clone"></a>}}
\if{latex}{\out{\hypertarget{method-prov4RLogger-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prov4RLogger$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
