% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProvAbstractObject.R
\docType{class}
\name{ProvAbstractObject}
\alias{ProvAbstractObject}
\title{PROV abstract class}
\format{
\code{ \link[R6]{R6Class}} object.
}
\value{
Object of \code{ \link[R6]{R6Class}} for modelling an PROV abstract Object
}
\description{
This class models an PROV abstract object
}
\details{
ProvAbstractObject
}
\note{
abstract class used internally by \pkg{prov4R}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{PROV}
\section{Super class}{
\code{\link[prov4R:prov4RLogger]{prov4R::prov4RLogger}} -> \code{ProvAbstractObject}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{wrap}}{wrapping XML element}

\item{\code{element}}{element}

\item{\code{namespace}}{namespace}

\item{\code{defaults}}{defaults}

\item{\code{attrs}}{attrs}

\item{\code{printAttrs}}{attrs to print}

\item{\code{parentAttrs}}{parent attrs}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ProvAbstractObject-new}{\code{ProvAbstractObject$new()}}
\item \href{#method-ProvAbstractObject-setIsDocument}{\code{ProvAbstractObject$setIsDocument()}}
\item \href{#method-ProvAbstractObject-isDocument}{\code{ProvAbstractObject$isDocument()}}
\item \href{#method-ProvAbstractObject-getRootElement}{\code{ProvAbstractObject$getRootElement()}}
\item \href{#method-ProvAbstractObject-getNamespace}{\code{ProvAbstractObject$getNamespace()}}
\item \href{#method-ProvAbstractObject-createElement}{\code{ProvAbstractObject$createElement()}}
\item \href{#method-ProvAbstractObject-addListElement}{\code{ProvAbstractObject$addListElement()}}
\item \href{#method-ProvAbstractObject-delListElement}{\code{ProvAbstractObject$delListElement()}}
\item \href{#method-ProvAbstractObject-contains}{\code{ProvAbstractObject$contains()}}
\item \href{#method-ProvAbstractObject-print}{\code{ProvAbstractObject$print()}}
\item \href{#method-ProvAbstractObject-decode}{\code{ProvAbstractObject$decode()}}
\item \href{#method-ProvAbstractObject-encode}{\code{ProvAbstractObject$encode()}}
\item \href{#method-ProvAbstractObject-validate}{\code{ProvAbstractObject$validate()}}
\item \href{#method-ProvAbstractObject-save}{\code{ProvAbstractObject$save()}}
\item \href{#method-ProvAbstractObject-isFieldInheritedFrom}{\code{ProvAbstractObject$isFieldInheritedFrom()}}
\item \href{#method-ProvAbstractObject-getClassName}{\code{ProvAbstractObject$getClassName()}}
\item \href{#method-ProvAbstractObject-getClass}{\code{ProvAbstractObject$getClass()}}
\item \href{#method-ProvAbstractObject-getNamespaceDefinition}{\code{ProvAbstractObject$getNamespaceDefinition()}}
\item \href{#method-ProvAbstractObject-getXmlElement}{\code{ProvAbstractObject$getXmlElement()}}
\item \href{#method-ProvAbstractObject-clone}{\code{ProvAbstractObject$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="prov4RLogger" data-id="ERROR"><a href='../../prov4R/html/prov4RLogger.html#method-prov4RLogger-ERROR'><code>prov4R::prov4RLogger$ERROR()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="prov4RLogger" data-id="INFO"><a href='../../prov4R/html/prov4RLogger.html#method-prov4RLogger-INFO'><code>prov4R::prov4RLogger$INFO()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="prov4RLogger" data-id="WARN"><a href='../../prov4R/html/prov4RLogger.html#method-prov4RLogger-WARN'><code>prov4R::prov4RLogger$WARN()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="prov4RLogger" data-id="logger"><a href='../../prov4R/html/prov4RLogger.html#method-prov4RLogger-logger'><code>prov4R::prov4RLogger$logger()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProvAbstractObject-new"></a>}}
\if{latex}{\out{\hypertarget{method-ProvAbstractObject-new}{}}}
\subsection{Method \code{new()}}{
Initializes an object of class \link{ProvAbstractObject}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvAbstractObject$new(
  xml = NULL,
  element = NULL,
  namespace = NULL,
  attrs = list(),
  defaults = list(),
  wrap = FALSE,
  logger = "INFO"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xml}}{object of class \link[XML]{XMLInternalNode-class}}

\item{\code{element}}{element}

\item{\code{namespace}}{namespace}

\item{\code{attrs}}{attrs}

\item{\code{defaults}}{defaults}

\item{\code{wrap}}{wrap}

\item{\code{logger}}{logger type}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProvAbstractObject-setIsDocument"></a>}}
\if{latex}{\out{\hypertarget{method-ProvAbstractObject-setIsDocument}{}}}
\subsection{Method \code{setIsDocument()}}{
Set if object is a document or not
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvAbstractObject$setIsDocument(isDocument)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{isDocument}}{object of class \code{logical}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProvAbstractObject-isDocument"></a>}}
\if{latex}{\out{\hypertarget{method-ProvAbstractObject-isDocument}{}}}
\subsection{Method \code{isDocument()}}{
Informs if the object is a document
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvAbstractObject$isDocument()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
object of class \code{logical}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProvAbstractObject-getRootElement"></a>}}
\if{latex}{\out{\hypertarget{method-ProvAbstractObject-getRootElement}{}}}
\subsection{Method \code{getRootElement()}}{
Get root XML element
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvAbstractObject$getRootElement()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
object of class \code{character}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProvAbstractObject-getNamespace"></a>}}
\if{latex}{\out{\hypertarget{method-ProvAbstractObject-getNamespace}{}}}
\subsection{Method \code{getNamespace()}}{
Get XML namespace
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvAbstractObject$getNamespace()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
object of class \code{character}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProvAbstractObject-createElement"></a>}}
\if{latex}{\out{\hypertarget{method-ProvAbstractObject-createElement}{}}}
\subsection{Method \code{createElement()}}{
Creates an element
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvAbstractObject$createElement(element, type = "text")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{element}}{element}

\item{\code{type}}{type. Default is "text"}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the typed element
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProvAbstractObject-addListElement"></a>}}
\if{latex}{\out{\hypertarget{method-ProvAbstractObject-addListElement}{}}}
\subsection{Method \code{addListElement()}}{
Add a metadata element to an element list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvAbstractObject$addListElement(field, metadataElement)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{field}}{field}

\item{\code{metadataElement}}{metadata element to add}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{TRUE} if added, \code{FALSE} otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProvAbstractObject-delListElement"></a>}}
\if{latex}{\out{\hypertarget{method-ProvAbstractObject-delListElement}{}}}
\subsection{Method \code{delListElement()}}{
Deletes a metadata element from an element list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvAbstractObject$delListElement(field, metadataElement)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{field}}{field}

\item{\code{metadataElement}}{metadata element to add}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{TRUE} if deleted, \code{FALSE} otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProvAbstractObject-contains"></a>}}
\if{latex}{\out{\hypertarget{method-ProvAbstractObject-contains}{}}}
\subsection{Method \code{contains()}}{
Indicates if an element list contains or not an element
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvAbstractObject$contains(field, metadataElement)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{field}}{field}

\item{\code{metadataElement}}{metadata element to add}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{TRUE} if contained, \code{FALSE} otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProvAbstractObject-print"></a>}}
\if{latex}{\out{\hypertarget{method-ProvAbstractObject-print}{}}}
\subsection{Method \code{print()}}{
Prints the element
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvAbstractObject$print(..., depth = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{any parameter to pass to print method}

\item{\code{depth}}{printing depth}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProvAbstractObject-decode"></a>}}
\if{latex}{\out{\hypertarget{method-ProvAbstractObject-decode}{}}}
\subsection{Method \code{decode()}}{
Decodes the object from an \pkg{XML} representation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvAbstractObject$decode(xml)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xml}}{object of class \link[XML]{XMLInternalNode-class} from \pkg{XML}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProvAbstractObject-encode"></a>}}
\if{latex}{\out{\hypertarget{method-ProvAbstractObject-encode}{}}}
\subsection{Method \code{encode()}}{
Encodes the object as XML
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvAbstractObject$encode(
  addNS = TRUE,
  validate = TRUE,
  strict = FALSE,
  encoding = "UTF-8"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{addNS}}{whether namespace has to be added. Default is \code{TRUE}}

\item{\code{validate}}{whether validation has to be done vs. XML schemas. Default is \code{TRUE}}

\item{\code{strict}}{whether strict validation has to be operated (raise an error if invalid). Default is \code{FALSE}}

\item{\code{encoding}}{encoding. Default is "UTF-8"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProvAbstractObject-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ProvAbstractObject-validate}{}}}
\subsection{Method \code{validate()}}{
Validates the object / XML vs. XML schemas
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvAbstractObject$validate(xml = NULL, strict = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xml}}{object of class \link[XML]{XMLInternalNode-class} from \pkg{XML}}

\item{\code{strict}}{strict validation or not}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{TRUE} if valid, \code{FALSE} otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProvAbstractObject-save"></a>}}
\if{latex}{\out{\hypertarget{method-ProvAbstractObject-save}{}}}
\subsection{Method \code{save()}}{
Saves the object as XML file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvAbstractObject$save(file, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{file name}

\item{\code{...}}{any parameter to pass to \code{encode()} method}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProvAbstractObject-isFieldInheritedFrom"></a>}}
\if{latex}{\out{\hypertarget{method-ProvAbstractObject-isFieldInheritedFrom}{}}}
\subsection{Method \code{isFieldInheritedFrom()}}{
Indicates the class from which field is inherited
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvAbstractObject$isFieldInheritedFrom(field)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{field}}{field}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
an object of class  \link[R6]{R6Class}, or \code{NULL}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProvAbstractObject-getClassName"></a>}}
\if{latex}{\out{\hypertarget{method-ProvAbstractObject-getClassName}{}}}
\subsection{Method \code{getClassName()}}{
Get class name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvAbstractObject$getClassName()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
object of class \code{character}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProvAbstractObject-getClass"></a>}}
\if{latex}{\out{\hypertarget{method-ProvAbstractObject-getClass}{}}}
\subsection{Method \code{getClass()}}{
Get class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvAbstractObject$getClass()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
object of class  \link[R6]{R6Class}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProvAbstractObject-getNamespaceDefinition"></a>}}
\if{latex}{\out{\hypertarget{method-ProvAbstractObject-getNamespaceDefinition}{}}}
\subsection{Method \code{getNamespaceDefinition()}}{
Get namespace definition
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvAbstractObject$getNamespaceDefinition(recursive = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{recursive}}{recursive}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a named \code{list} of the XML namespaces
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProvAbstractObject-getXmlElement"></a>}}
\if{latex}{\out{\hypertarget{method-ProvAbstractObject-getXmlElement}{}}}
\subsection{Method \code{getXmlElement()}}{
Get XML element name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvAbstractObject$getXmlElement()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
object of class \code{character}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProvAbstractObject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ProvAbstractObject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvAbstractObject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
