% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProvDocument.R
\docType{class}
\name{ProvDocument}
\alias{ProvDocument}
\title{PROV document class}
\format{
\code{ \link[R6]{R6Class}} object.
}
\value{
Object of \code{ \link[R6]{R6Class}} for modelling an PROV document
}
\description{
This class models an PROV document
}
\details{
ProvDocument
}
\examples{
  #encoding
  prov <- ProvDocument$new()
  prov$encode()

}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{atom}
\keyword{feed}
\section{Super classes}{
\code{\link[prov4R:prov4RLogger]{prov4R::prov4RLogger}} -> \code{\link[prov4R:ProvAbstractObject]{prov4R::ProvAbstractObject}} -> \code{ProvDocument}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{entity}}{entity}

\item{\code{activity}}{activity}

\item{\code{wasGeneratedBy}}{wasGeneratedBy}

\item{\code{agent}}{agent}

\item{\code{person}}{person}

\item{\code{organization}}{organization}

\item{\code{softwareAgent}}{softwareAgent}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ProvDocument-new}{\code{ProvDocument$new()}}
\item \href{#method-ProvDocument-addEntity}{\code{ProvDocument$addEntity()}}
\item \href{#method-ProvDocument-addActivity}{\code{ProvDocument$addActivity()}}
\item \href{#method-ProvDocument-addGeneration}{\code{ProvDocument$addGeneration()}}
\item \href{#method-ProvDocument-addUsage}{\code{ProvDocument$addUsage()}}
\item \href{#method-ProvDocument-addAgent}{\code{ProvDocument$addAgent()}}
\item \href{#method-ProvDocument-addPerson}{\code{ProvDocument$addPerson()}}
\item \href{#method-ProvDocument-addOrganization}{\code{ProvDocument$addOrganization()}}
\item \href{#method-ProvDocument-addSoftwareAgent}{\code{ProvDocument$addSoftwareAgent()}}
\item \href{#method-ProvDocument-clone}{\code{ProvDocument$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="prov4RLogger" data-id="ERROR"><a href='../../prov4R/html/prov4RLogger.html#method-prov4RLogger-ERROR'><code>prov4R::prov4RLogger$ERROR()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="prov4RLogger" data-id="INFO"><a href='../../prov4R/html/prov4RLogger.html#method-prov4RLogger-INFO'><code>prov4R::prov4RLogger$INFO()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="prov4RLogger" data-id="WARN"><a href='../../prov4R/html/prov4RLogger.html#method-prov4RLogger-WARN'><code>prov4R::prov4RLogger$WARN()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="prov4RLogger" data-id="logger"><a href='../../prov4R/html/prov4RLogger.html#method-prov4RLogger-logger'><code>prov4R::prov4RLogger$logger()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="addListElement"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-addListElement'><code>prov4R::ProvAbstractObject$addListElement()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="contains"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-contains'><code>prov4R::ProvAbstractObject$contains()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="createElement"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-createElement'><code>prov4R::ProvAbstractObject$createElement()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="decode"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-decode'><code>prov4R::ProvAbstractObject$decode()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="delListElement"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-delListElement'><code>prov4R::ProvAbstractObject$delListElement()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="encode"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-encode'><code>prov4R::ProvAbstractObject$encode()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="getClass"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-getClass'><code>prov4R::ProvAbstractObject$getClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="getClassName"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-getClassName'><code>prov4R::ProvAbstractObject$getClassName()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="getId"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-getId'><code>prov4R::ProvAbstractObject$getId()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="getNamespace"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-getNamespace'><code>prov4R::ProvAbstractObject$getNamespace()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="getNamespaceDefinition"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-getNamespaceDefinition'><code>prov4R::ProvAbstractObject$getNamespaceDefinition()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="getRef"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-getRef'><code>prov4R::ProvAbstractObject$getRef()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="getRootElement"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-getRootElement'><code>prov4R::ProvAbstractObject$getRootElement()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="getXmlElement"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-getXmlElement'><code>prov4R::ProvAbstractObject$getXmlElement()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="isDocument"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-isDocument'><code>prov4R::ProvAbstractObject$isDocument()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="isFieldInheritedFrom"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-isFieldInheritedFrom'><code>prov4R::ProvAbstractObject$isFieldInheritedFrom()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="print"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-print'><code>prov4R::ProvAbstractObject$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="save"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-save'><code>prov4R::ProvAbstractObject$save()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="setId"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-setId'><code>prov4R::ProvAbstractObject$setId()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="setIsDocument"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-setIsDocument'><code>prov4R::ProvAbstractObject$setIsDocument()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="setProvAttr"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-setProvAttr'><code>prov4R::ProvAbstractObject$setProvAttr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="setRef"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-setRef'><code>prov4R::ProvAbstractObject$setRef()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="validate"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-validate'><code>prov4R::ProvAbstractObject$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProvDocument-new"></a>}}
\if{latex}{\out{\hypertarget{method-ProvDocument-new}{}}}
\subsection{Method \code{new()}}{
Initializes a \link{ProvDocument}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvDocument$new(xml = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xml}}{object of class \link[XML]{XMLInternalNode-class} from \pkg{XML}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProvDocument-addEntity"></a>}}
\if{latex}{\out{\hypertarget{method-ProvDocument-addEntity}{}}}
\subsection{Method \code{addEntity()}}{
Adds entity
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvDocument$addEntity(entity)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{entity}}{entity}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProvDocument-addActivity"></a>}}
\if{latex}{\out{\hypertarget{method-ProvDocument-addActivity}{}}}
\subsection{Method \code{addActivity()}}{
Adds activity
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvDocument$addActivity(activity)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{activity}}{activity}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProvDocument-addGeneration"></a>}}
\if{latex}{\out{\hypertarget{method-ProvDocument-addGeneration}{}}}
\subsection{Method \code{addGeneration()}}{
Adds generation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvDocument$addGeneration(entity, activity, time)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{entity}}{entity}

\item{\code{activity}}{activity}

\item{\code{time}}{time}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProvDocument-addUsage"></a>}}
\if{latex}{\out{\hypertarget{method-ProvDocument-addUsage}{}}}
\subsection{Method \code{addUsage()}}{
Adds usage
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvDocument$addUsage(entity, activity, time)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{entity}}{entity}

\item{\code{activity}}{activity}

\item{\code{time}}{time}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProvDocument-addAgent"></a>}}
\if{latex}{\out{\hypertarget{method-ProvDocument-addAgent}{}}}
\subsection{Method \code{addAgent()}}{
Adds agent
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvDocument$addAgent(agent)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{agent}}{agent}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProvDocument-addPerson"></a>}}
\if{latex}{\out{\hypertarget{method-ProvDocument-addPerson}{}}}
\subsection{Method \code{addPerson()}}{
Adds person
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvDocument$addPerson(person)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{person}}{person}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProvDocument-addOrganization"></a>}}
\if{latex}{\out{\hypertarget{method-ProvDocument-addOrganization}{}}}
\subsection{Method \code{addOrganization()}}{
Adds organization
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvDocument$addOrganization(organization)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{organization}}{organization}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProvDocument-addSoftwareAgent"></a>}}
\if{latex}{\out{\hypertarget{method-ProvDocument-addSoftwareAgent}{}}}
\subsection{Method \code{addSoftwareAgent()}}{
Adds software agent
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvDocument$addSoftwareAgent(softwareAgent)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{softwareAgent}}{softwareAgent}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProvDocument-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ProvDocument-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvDocument$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
