% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProvGeneration.R
\docType{class}
\name{ProvGeneration}
\alias{ProvGeneration}
\title{PROV generation class}
\format{
\code{ \link[R6]{R6Class}} object.
}
\value{
Object of \code{ \link[R6]{R6Class}} for modelling an PROV generation
}
\description{
This class models an PROV generation
}
\details{
ProvGeneration
}
\examples{
  #encoding
  gen <- ProvGeneration$new()
  gen$encode()

}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{atom}
\keyword{feed}
\section{Super classes}{
\code{\link[prov4R:prov4RLogger]{prov4R::prov4RLogger}} -> \code{\link[prov4R:ProvAbstractObject]{prov4R::ProvAbstractObject}} -> \code{ProvGeneration}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{entity}}{entity}

\item{\code{activity}}{activity}

\item{\code{time}}{time}

\item{\code{label}}{label}

\item{\code{location}}{location}

\item{\code{type}}{type}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ProvGeneration-new}{\code{ProvGeneration$new()}}
\item \href{#method-ProvGeneration-setEntity}{\code{ProvGeneration$setEntity()}}
\item \href{#method-ProvGeneration-setActivity}{\code{ProvGeneration$setActivity()}}
\item \href{#method-ProvGeneration-setTime}{\code{ProvGeneration$setTime()}}
\item \href{#method-ProvGeneration-addLabel}{\code{ProvGeneration$addLabel()}}
\item \href{#method-ProvGeneration-addLocation}{\code{ProvGeneration$addLocation()}}
\item \href{#method-ProvGeneration-addType}{\code{ProvGeneration$addType()}}
\item \href{#method-ProvGeneration-clone}{\code{ProvGeneration$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="prov4RLogger" data-id="ERROR"><a href='../../prov4R/html/prov4RLogger.html#method-prov4RLogger-ERROR'><code>prov4R::prov4RLogger$ERROR()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="prov4RLogger" data-id="INFO"><a href='../../prov4R/html/prov4RLogger.html#method-prov4RLogger-INFO'><code>prov4R::prov4RLogger$INFO()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="prov4RLogger" data-id="WARN"><a href='../../prov4R/html/prov4RLogger.html#method-prov4RLogger-WARN'><code>prov4R::prov4RLogger$WARN()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="prov4RLogger" data-id="logger"><a href='../../prov4R/html/prov4RLogger.html#method-prov4RLogger-logger'><code>prov4R::prov4RLogger$logger()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="addListElement"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-addListElement'><code>prov4R::ProvAbstractObject$addListElement()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="contains"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-contains'><code>prov4R::ProvAbstractObject$contains()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="createElement"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-createElement'><code>prov4R::ProvAbstractObject$createElement()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="decode"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-decode'><code>prov4R::ProvAbstractObject$decode()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="delListElement"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-delListElement'><code>prov4R::ProvAbstractObject$delListElement()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="encode"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-encode'><code>prov4R::ProvAbstractObject$encode()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="getClass"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-getClass'><code>prov4R::ProvAbstractObject$getClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="getClassName"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-getClassName'><code>prov4R::ProvAbstractObject$getClassName()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="getId"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-getId'><code>prov4R::ProvAbstractObject$getId()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="getNamespace"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-getNamespace'><code>prov4R::ProvAbstractObject$getNamespace()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="getNamespaceDefinition"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-getNamespaceDefinition'><code>prov4R::ProvAbstractObject$getNamespaceDefinition()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="getRef"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-getRef'><code>prov4R::ProvAbstractObject$getRef()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="getRootElement"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-getRootElement'><code>prov4R::ProvAbstractObject$getRootElement()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="getXmlElement"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-getXmlElement'><code>prov4R::ProvAbstractObject$getXmlElement()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="isDocument"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-isDocument'><code>prov4R::ProvAbstractObject$isDocument()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="isFieldInheritedFrom"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-isFieldInheritedFrom'><code>prov4R::ProvAbstractObject$isFieldInheritedFrom()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="print"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-print'><code>prov4R::ProvAbstractObject$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="save"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-save'><code>prov4R::ProvAbstractObject$save()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="setId"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-setId'><code>prov4R::ProvAbstractObject$setId()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="setIsDocument"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-setIsDocument'><code>prov4R::ProvAbstractObject$setIsDocument()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="setProvAttr"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-setProvAttr'><code>prov4R::ProvAbstractObject$setProvAttr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="setRef"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-setRef'><code>prov4R::ProvAbstractObject$setRef()</code></a></span></li>
<li><span class="pkg-link" data-pkg="prov4R" data-topic="ProvAbstractObject" data-id="validate"><a href='../../prov4R/html/ProvAbstractObject.html#method-ProvAbstractObject-validate'><code>prov4R::ProvAbstractObject$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProvGeneration-new"></a>}}
\if{latex}{\out{\hypertarget{method-ProvGeneration-new}{}}}
\subsection{Method \code{new()}}{
Initializes a \link{ProvGeneration}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvGeneration$new(xml = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xml}}{object of class \link[XML]{XMLInternalNode-class} from \pkg{XML}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProvGeneration-setEntity"></a>}}
\if{latex}{\out{\hypertarget{method-ProvGeneration-setEntity}{}}}
\subsection{Method \code{setEntity()}}{
Set entity
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvGeneration$setEntity(entity)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{entity}}{entity}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProvGeneration-setActivity"></a>}}
\if{latex}{\out{\hypertarget{method-ProvGeneration-setActivity}{}}}
\subsection{Method \code{setActivity()}}{
Set activity
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvGeneration$setActivity(activity)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{activity}}{activity}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProvGeneration-setTime"></a>}}
\if{latex}{\out{\hypertarget{method-ProvGeneration-setTime}{}}}
\subsection{Method \code{setTime()}}{
Set time
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvGeneration$setTime(time)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time}}{time}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProvGeneration-addLabel"></a>}}
\if{latex}{\out{\hypertarget{method-ProvGeneration-addLabel}{}}}
\subsection{Method \code{addLabel()}}{
Adds label
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvGeneration$addLabel(label)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{label}}{label}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{TRUE} if added, \code{FALSE} otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProvGeneration-addLocation"></a>}}
\if{latex}{\out{\hypertarget{method-ProvGeneration-addLocation}{}}}
\subsection{Method \code{addLocation()}}{
Adds location
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvGeneration$addLocation(location)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{location}}{location}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{TRUE} if added, \code{FALSE} otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProvGeneration-addType"></a>}}
\if{latex}{\out{\hypertarget{method-ProvGeneration-addType}{}}}
\subsection{Method \code{addType()}}{
Adds type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvGeneration$addType(type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{type}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{TRUE} if added, \code{FALSE} otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProvGeneration-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ProvGeneration-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvGeneration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
