% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProvNamespace.R
\docType{class}
\name{ProvNamespace}
\alias{ProvNamespace}
\title{ProvNamespace}
\format{
\code{ \link[R6]{R6Class}} object.
}
\value{
Object of \code{ \link[R6]{R6Class}} for modelling an Atom Namespace
}
\description{
ProvNamespace

ProvNamespace
}
\note{
ISO class used internally by prov4R for specifying XML namespaces
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{metadata}
\keyword{namespace}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{id}

\item{\code{uri}}{uri}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ProvNamespace-new}{\code{ProvNamespace$new()}}
\item \href{#method-ProvNamespace-getDefinition}{\code{ProvNamespace$getDefinition()}}
\item \href{#method-ProvNamespace-clone}{\code{ProvNamespace$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProvNamespace-new"></a>}}
\if{latex}{\out{\hypertarget{method-ProvNamespace-new}{}}}
\subsection{Method \code{new()}}{
Initializes an \link{ProvNamespace}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvNamespace$new(id, uri)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{uri}}{uri}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProvNamespace-getDefinition"></a>}}
\if{latex}{\out{\hypertarget{method-ProvNamespace-getDefinition}{}}}
\subsection{Method \code{getDefinition()}}{
Get definition
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvNamespace$getDefinition()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a named list defining the namespace
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProvNamespace-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ProvNamespace-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvNamespace$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
