% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProvNamespace.R
\name{registerProvNamespace}
\alias{registerProvNamespace}
\title{registerProvNamespace}
\usage{
registerProvNamespace(id, uri, force)
}
\arguments{
\item{id}{prefix of the namespace}

\item{uri}{URI of the namespace}

\item{force}{logical parameter indicating if registration has be to be forced
in case the identified namespace is already registered}
}
\description{
\code{registerProvNamespace} allows to register a new namespace
in \pkg{prov4R}
}
\examples{
  registerProvNamespace(id = "myprefix", uri = "http://someuri")

}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
