% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProvSchemas.R
\name{registerProvSchema}
\alias{registerProvSchema}
\title{registerProvSchema}
\usage{
registerProvSchema(xsdFile)
}
\arguments{
\item{xsdFile}{the schema XSD file}
}
\description{
\code{registerProvSchema} allows to register a new schema
in \pkg{prov4R}
}
\examples{
  atom_xsd_file <- system.file("extdata/schemas/atom/atom.xsd", package = "prov4R")
  registerProvSchema(xsdFile = atom_xsd_file)

}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
